orchestrator/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
httpx==0.27.*
qdrant-client==1.9.*
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
# Optional (dev/test/quality)
pytest>=8.2
pytest-asyncio>=0.23
black>=24.3
ruff>=0.4
mypy>=1.9
types-requests>=2.32
bandit==1.7.9

python-multipart==0.0.9
watchfiles==0.22.0
PyYAML==6.0.2

# (Se usi diff/patch lato py per utilità varie)
unidiff==0.7.5

# Embedded tools (usati via subprocess, ma li teniamo anche come dipendenze python-side dove utile)
isort==5.13.2
tree_sitter>=0.21.0
tree_sitter_languages>=1.10.2

